// src/app/foundation/core/models/project-config.generated.ts
export interface ConfigInputDomain {
  id: string;
  label: string;
  enabled?: boolean;
  modules: ConfigInputModule[];
}
export interface ConfigInputModule {
  id: string;
  label: string;
  icon?: string;
  apiUrl: string;
  fallbackJsonAr?: string;
  fallbackJsonEn?: string;
  idField?: string;
  actions?: {
    create?: boolean;
    edit?: boolean;
    view?: boolean;
    delete?: boolean;
  };
  enabled?: boolean;
  subModules?: ConfigInputSubModule[];
}
export interface ConfigInputSubModule {
  id: string;
  label: string;
  icon?: string;
  /**
   * مثال: /org/communities
   */
  path: string;
  enabled?: boolean;
}